Struct strymon_coordinator::catalog::Catalog

pub struct Catalog { /* fields omitted */ }

A representation of the current catalog state.

Methods

impl Catalog

pub fn new(addr: (String, u16)) -> Self

Creates a new catalog instance, containing a self-referring topic.

The created instance will contain a single topic announcing the address of the catalog itself (given through the addr argument). The created initial topic will have the name "$catalog" and is of schema Service(CatalogRPC).

pub fn add_executor(&mut self, executor: Executor)

Inserts a new available executor into the catalog.

pub fn remove_executor(&mut self, id: ExecutorId)

Removes an existing executor from the catalog.

Important traits for Executors<'a>
impl<'a> Iterator for Executors<'a> type Item = &'a Executor;

pub fn executors<'a>(&'a self) -> Executors<'a>

Returns an iterator over all existing executors.

pub fn add_job(&mut self, job: Job)

Adds a new completely spawned job.

pub fn remove_job(&mut self, id: JobId)

Removes a job from the catalog.

pub fn publish(
&mut self,
job: JobId,
name: String,
addr: (String, u16),
schema: TopicSchema
) -> Result<Topic, PublishError>

Creates and announces a new topic.

This fails if a topic with the same name already exists.

pub fn unpublish(
&mut self,
job_id: JobId,
topic: TopicId
) -> Result<(), UnpublishError>

Removes a publication from the catalog.

pub fn lookup(&self, name: &str) -> Option<Topic>

Finds a topic by name.

Returns None if the topic does not exist.

pub fn subscribe(&mut self, job_id: JobId, topic: TopicId)

Adds a new subscription to the catalog.

pub fn unsubscribe(
&mut self,
job_id: JobId,
topic: TopicId
) -> Result<(), UnsubscribeError>

Removes a subscription from the catalog.

pub fn request(&self, req: RequestBuf<CatalogRPC>) -> Result<()>

Decodes and responds to an incoming catalog service request.