Module strymon_coordinator::handler

The coordinator request handler.

The Coordinator type implements most of the coordinator logic. The Coordinator stores some additional state not exposed in the catalog, such as connection handles or partially spawned jobs. Clients submit their requests through a CoordinatorRef, which tracks client-associated state.

Structs

Coordinator
The coordinator instance, contains the current system state and handles incoming requests.

CoordinatorRef
A cloneable reference to the coordinator instance, tracking any state created by this handle.